+++ src/BusinessLogic/BootstrapComponent.php
@@ -21,6 +21,7 @@ use Packlink\BusinessLogic\Language\TranslationService;
 use Packlink\BusinessLogic\Location\LocationService;
 use Packlink\BusinessLogic\Order\OrderService;
 use Packlink\BusinessLogic\OrderShipmentDetails\OrderShipmentDetailsService;
+use Packlink\BusinessLogic\PostalCode\PostalCodeTransformer;
 use Packlink\BusinessLogic\Registration\RegistrationLegalPolicy;
 use Packlink\BusinessLogic\Registration\RegistrationRequest;
 use Packlink\BusinessLogic\Registration\RegistrationService;
@@ -172,6 +173,13 @@ class BootstrapComponent extends \Logeecom\Infrastructure\BootstrapComponent
             AutoTestService::CLASS_NAME,
             function () {
                 return new AutoTestService();
+            }
+        );
+
+        ServiceRegister::registerService(
+            PostalCodeTransformer::CLASS_NAME,
+            function () {
+                return new PostalCodeTransformer();
             }
         );
     }