+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Added postal code transformer service that transforms postal code into supported postal code format for some countries.
+
 ### Changed
 - Changed how the default parcel is validated.